pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "role-forge"
version = "0.0.1"
description = "Canonical role-definition toolkit for coding agents"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Zhan Rongrui", email = "2742392377@qq.com" }]
dependencies = [
  "loguru>=0.7.0",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "typer>=0.15.0",
]

[project.scripts]
agent-caster = "role_forge.legacy_cli:main"
role-forge = "role_forge.cli:app"

[project.entry-points."role_forge.adapters"]
opencode = "role_forge.adapters.opencode:OpenCodeAdapter"
claude = "role_forge.adapters.claude:ClaudeAdapter"
cursor = "role_forge.adapters.cursor:CursorAdapter"

[project.urls]
Homepage = "https://github.com/zrr1999/role-forge"
Repository = "https://github.com/zrr1999/role-forge"

[dependency-groups]
dev = [
  { include-group = "docs" },
  { include-group = "test" },
]
docs = [
  "zensical>=0.0.24",
]
test = [
  "pytest>=8.0",
  "pytest-cov>=6.0",
  "syrupy>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--strict-markers", "--tb=short"]

[tool.coverage.run]
branch = true
source = ["role_forge"]
omit = ["*/tests/*"]

[tool.coverage.report]
precision = 2
show_missing = true
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "@abstractmethod",
  "\\.\\.\\.",
]

[tool.coverage.xml]
output = "coverage.xml"

[tool.ruff]
line-length = 100
src = ["src", "tests"]
target-version = "py312"
exclude = ["src/role_forge/_version.py"]

[tool.ruff.lint]
select = [
  "F",  # pyflakes
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "I",  # isort
  "B",  # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
  "SIM",  # flake8-simplify
  "RUF",  # ruff-specific rules
  "PGH",  # pygrep-hooks
  "FA",  # flake8-future-annotations
  "YTT",  # flake8-2020
]
ignore = ["PGH003", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["role_forge"]
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "I002"]
"tests/**" = ["B011", "I002"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/role_forge"]

[tool.uv]
default-groups = ["dev"]
